Change id field from Integer to Long

+++ src/main/java/org/boutwaretech/weightspec/repositories/BodyMeasurementApprovalRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "approvals", path = "approvals")
-public interface BodyMeasurementApprovalRepository extends PagingAndSortingRepository<BodyMeasurementApproval, Integer> {
+public interface BodyMeasurementApprovalRepository extends PagingAndSortingRepository<BodyMeasurementApproval, Long> {
 
 }
